@@ -18,28 +18,29 @@
   "packageManager": "yarn@3.2.0",
   "devDependencies": {
     "@nrwl/cli": "13.9.4",
-    "@nrwl/eslint-plugin-nx": "13.9.4",
-    "@nrwl/jest": "13.9.4",
-    "@nrwl/js": "13.9.4",
-    "@nrwl/linter": "13.9.4",
-    "@nrwl/workspace": "13.9.4",
+    "@nrwl/eslint-plugin-nx": "15.2.1",
+    "@nrwl/jest": "15.2.1",
+    "@nrwl/js": "15.2.1",
+    "@nrwl/linter": "15.2.1",
+    "@nrwl/workspace": "15.2.1",
     "@types/glob": "7.2.0",
-    "@types/jest": "27.0.2",
+    "@types/jest": "28.1.8",
     "@types/node": "16.11.7",
     "@types/tmp": "0.2.3",
-    "@typescript-eslint/eslint-plugin": "~5.10.0",
-    "@typescript-eslint/parser": "~5.10.0",
-    "eslint": "~8.7.0",
+    "@typescript-eslint/eslint-plugin": "5.44.0",
+    "@typescript-eslint/parser": "5.44.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-import-resolver-typescript": "2.7.0",
     "eslint-plugin-import": "2.25.4",
-    "jest": "27.2.3",
+    "jest": "28.1.3",
     "nano-staged": "0.6.0",
-    "nx": "13.9.4",
+    "nx": "15.2.1",
     "simple-git-hooks": "2.7.0",
     "syncpack": "6.2.0",
-    "ts-jest": "27.0.5",
-    "typescript": "~4.5.2"
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
+    "typescript": "4.8.4"
   },
   "dependencies": {
     "@babel/core": "^7.10.4",
@@ -52,15 +53,12 @@
     "find-up": "^5.0.0",
     "glob": "^7.1.2",
     "gzip-size": "^6.0.0",
-    "prettier": "^2.5.1",
+    "prettier": "^2.8.0",
     "pretty-bytes": "^5.6.0",
     "terser": "^5.5.1",
     "tslib": "^2.3.1",
     "webpack": "^5.21.2",
     "workspace-tools": "^0.16.2",
     "yargs": "^13.3.2"
-  },
-  "resolutions": {
-    "@nrwl/js@13.9.4": "patch:@nrwl/js@npm:13.9.4#.yarn/patches/@nrwl-js-npm-13.9.4-4c81228b07.patch"
   }
 }
